add "complete" param for non-shallow fetches
Shallow fetches are more efficient, but sometimes you need all of the
history. Setting a sufficiently large depth is not a good workaround for 2 reasons:

1. Setting a large depth like 2345678 (which, let's say, is
   effectively infinity) in the YAML actually fails with the following
   error:

   ```
   panic: Unable to unarmshal vargs. json: cannot unmarshal number 2.345678e+06 into Go value of type int
   ```

2. All shallow clones require the usage of a temporary scratch
   file/dir (in cgit), and they can't use the fast-path that full-history
   clones use.

The reason you sometimes need all of the history is if you want to
build a commit that's very old. For example, a commit that's the
1000th-old commit on a branch would require a depth of 1000, but it's
unlikely that the builder would know how many commits ago the commit
was. So, the builder would need to avoid using shallow clones
altogether.