Allow cloning using specified SSH key
The motivation behind this commit is that we wish to administer Drone
using a single "machine user" that has access to a restricted set of
repositories rather than our individual GitHub user accounts that
require setting permissions that are more permissive than we are
comfortable with.

This allows passing in the "DRONE_PRIVATE_SSH_KEY" environment variable
at clone time, which can be configured using the "secrets" parameter.
The repository would have to be configured to fetch using `git@github`
which can also be done by passing in an enviornment variable.

If the environment variable is not detected, then it falls back to the
standard method of using `.netrc`.